+++ src/modules/playersSkills/components/skillsTableOptionsModal/__tests__/skillsTableOptions.test.tsx
@@ -3,12 +3,18 @@ import 'reflect-metadata';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import renderer, { ReactTestRendererJSON } from 'react-test-renderer';
+import { container } from 'tsyringe';
+
+import { PlayersSkillsFactoryMock } from '@src/__mocks__/playersSkillsFactory.mock';
+import { PlayersSkillsFactory } from '@src/modules/playersSkills/services/playersSkills.factory';
 
 import SkillsTableOptions from '../skillsTableOptions.component';
 
 describe('SkillsTableOptions', () => {
     const oldCreatePortal = ReactDOM.createPortal;
     beforeAll(() => {
+        container.register(PlayersSkillsFactory, PlayersSkillsFactoryMock);
+
         (ReactDOM.createPortal as unknown) = jest.fn((element, _node) => {
             return element;
         });
