+++ src/modules/playersSkills/services/playersSkills.factory.ts
@@ -1,9 +1,11 @@
 import { container, singleton } from 'tsyringe';
 
-import { CheckboxItem } from '@src/common/components/checkBoxOptions/checkboxItem.model';
-import { _teamPlayersSkillsSettingsKey, _transferListSettingsKey } from '@src/common/services/settings/settings';
-
 import { PlayersSkillsTableService } from './playersSkillsTable.service';
+import {
+    PageSettingsService,
+    TeamPlayersSkillsSettingsService,
+    TransferListSettingsService,
+} from './playersSkillsViewSettings';
 import { TeamPlayersSkillsTableService } from './teamPlayersSkillsTable.service';
 import { TransferListSkillsTableService } from './transferListSkillsTable.service';
 
@@ -17,36 +19,6 @@ export function getPlayersSkillsPageType(): PlayersSkillsPages | undefined {
     if (url.indexOf('plantel.asp') > -1 && url.indexOf('filtro=1') > -1) {
         return 'team-players-list-skills';
     }
-}
-
-export interface PlayersSkillsViewSettings {
-    hiddenColumns: string[];
-}
-
-export interface PageSettingsService {
-    key: string;
-    columnsConfig: CheckboxItem[];
-}
-
-@singleton()
-export class TransferListSettingsService implements PageSettingsService {
-    key = _transferListSettingsKey;
-    columnsConfig: CheckboxItem[] = [
-        { name: 'country', isChecked: true, label: 'Country' },
-        { name: 'wage', isChecked: true, label: 'Wage' },
-        { name: 'gs', isChecked: true, label: 'GS' },
-        { name: 'gsFutureMin', isChecked: true, label: 'GS future min' },
-        { name: 'gsFutureMax', isChecked: true, label: 'GS future max' },
-    ];
-}
-
-@singleton()
-export class TeamPlayersSkillsSettingsService implements PageSettingsService {
-    key = _teamPlayersSkillsSettingsKey;
-    columnsConfig: CheckboxItem[] = [
-        { name: 'country', isChecked: true, label: 'Country' },
-        { name: 'gs', isChecked: true, label: 'GS' },
-    ];
 }
 
 @singleton()
